separate login service from message stream service

+++ app/src/main/java/com/example/grpc_chat_android/view/fragments/LoginFragment.kt
@@ -5,15 +5,14 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.view.inputmethod.EditorInfo
+import android.widget.Toast
 import androidx.datastore.preferences.core.Preferences
 import androidx.fragment.app.Fragment
-import androidx.fragment.app.activityViewModels
 import androidx.fragment.app.viewModels
 import androidx.lifecycle.lifecycleScope
 import androidx.navigation.fragment.findNavController
 import com.example.grpc_chat_android.databinding.FragmentLoginBinding
 import com.example.grpc_chat_android.models.Chat
-import com.example.grpc_chat_android.viewmodel.MainActivityViewModel
 import com.example.grpc_chat_android.viewmodel.SignInViewModel
 import dagger.hilt.android.AndroidEntryPoint
 import javax.inject.Inject
@@ -26,7 +25,6 @@ class LoginFragment : Fragment() {
         get() = _binding!!
 
     private val viewModel: SignInViewModel by viewModels()
-    private val activityViewModel: MainActivityViewModel by activityViewModels()
 
     @Inject
     lateinit var key: Preferences.Key<String>
@@ -54,26 +52,35 @@ class LoginFragment : Fragment() {
         }
 
         binding.btLogin.setOnClickListener {
-            activityViewModel.login(
+            viewModel.login(
                 Chat.LoginRequest.newBuilder()
                     .setPhonenumber(binding.loginEtPhone.text.toString())
                     .setPassword(binding.loginEtPassword.text.toString()).build()
             )
-            lifecycleScope.launch {
-                viewModel.saveUserPhone(
-                    binding.loginEtPhone.text.toString(),
-                    requireContext(),
-                    key
-                )
-                findNavController()
-                    .navigate(
-                        LoginFragmentDirections.actionLoginFragmentToChatListFragment(
-                            binding.loginEtPhone.text.toString()
-                        )
-                    )
-            }
-            viewModel.deleteAll()
         }
+
+        viewModel.message.observe(viewLifecycleOwner, {
+            // Temporary workaround with string matching
+            when (it) {
+                "Login Successful" -> {
+                    Toast.makeText(requireContext(), it, Toast.LENGTH_SHORT).show()
+                    lifecycleScope.launch {
+                        viewModel.saveUserPhone(
+                            binding.loginEtPhone.text.toString(),
+                            requireContext(),
+                            key
+                        )
+                        findNavController().navigate(
+                            LoginFragmentDirections.actionLoginFragmentToChatListFragment(
+                                binding.loginEtPhone.text.toString()
+                            )
+                        )
+                    }
+                    viewModel.deleteAll()
+                }
+                else -> Toast.makeText(requireContext(), it, Toast.LENGTH_SHORT).show()
+            }
+        })
         return binding.root
     }
